ai_agent_python/test_simple.py
import google.generativeai as genai
import os
from dotenv import load_dotenv

load_dotenv()
api_key = os.getenv("GEMINI_API_KEY")

genai.configure(api_key=api_key)
model = genai.GenerativeModel('gemini-2.0-flash')
try:
    response = model.generate_content("Say hello in 2 words")
    print(response.text)
except Exception as e:
    print(f"Error: {e}")
