ai_agent_python/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
httpx==0.27.0

google-generativeai
python-dotenv
yfinance
